Markdown Parser Rework
