+++ quartz/plugins/transformers/html/syntax.ts
@@ -1,4 +1,4 @@
-import { QuartzTransformerPlugin } from "../types"
+import { HtmlTransformerPlugin, QuartzTransformerPlugin } from "../../types"
 import rehypePrettyCode, { Options as CodeOptions, Theme as CodeTheme } from "rehype-pretty-code"
 
 interface Theme extends Record<string, CodeTheme> {
@@ -19,12 +19,12 @@ const defaultOptions: Options = {
   keepBackground: false,
 }
 
-export const SyntaxHighlighting: QuartzTransformerPlugin<Partial<Options>> = (userOpts) => {
+export const SyntaxHighlighting: HtmlTransformerPlugin<Partial<Options>> = (userOpts) => {
   const opts: CodeOptions = { ...defaultOptions, ...userOpts }
 
   return {
     name: "SyntaxHighlighting",
-    htmlPlugins() {
+    transformation() {
       return [[rehypePrettyCode, opts]]
     },
   }
