--- quartz/plugins/transformers/linebreaks.ts
@@ -1,11 +1,0 @@
-import { QuartzTransformerPlugin } from "../types"
-import remarkBreaks from "remark-breaks"
-
-export const HardLineBreaks: QuartzTransformerPlugin = () => {
-  return {
-    name: "HardLineBreaks",
-    markdownPlugins() {
-      return [remarkBreaks]
-    },
-  }
-}
