+++ quartz/plugins/transformers/markdown/latex.ts
@@ -1,0 +1,11 @@
+import remarkMath from "remark-math"
+import { MarkdownTransformerPlugin } from "../../types"
+
+export const Latex: MarkdownTransformerPlugin = () => {
+  return {
+    name: "Latex",
+    transformation() {
+      return [remarkMath]
+    },
+  }
+}
