+++ quartz/plugins/transformers/markdown/linebreaks.ts
@@ -1,0 +1,11 @@
+import { MarkdownTransformerPlugin } from "../../types"
+import remarkBreaks from "remark-breaks"
+
+export const HardLineBreaks: MarkdownTransformerPlugin = () => {
+  return {
+    name: "HardLineBreaks",
+    transformation() {
+      return [remarkBreaks]
+    },
+  }
+}
