+++ quartz/plugins/transformers/markdown/roamSpecialEmbeds.ts
@@ -1,54 +1,22 @@
-import { QuartzTransformerPlugin } from "../types"
+import { MarkdownTransformerPlugin } from "../../types"
 import { PluggableList } from "unified"
-import { SKIP, visit } from "unist-util-visit"
+import { visit } from "unist-util-visit"
 import { ReplaceFunction, findAndReplace as mdastFindReplace } from "mdast-util-find-and-replace"
 import { Root, Html, Paragraph, Text, Link, Parent } from "mdast"
-import { Node } from "unist"
 import { VFile } from "vfile"
 import { BuildVisitor } from "unist-util-visit"
 
 export interface Options {
-  orComponent: boolean
-  TODOComponent: boolean
-  DONEComponent: boolean
   videoComponent: boolean
   audioComponent: boolean
   pdfComponent: boolean
-  blockquoteComponent: boolean
-  tableComponent: boolean
-  attributeComponent: boolean
 }
 
 const defaultOptions: Options = {
-  orComponent: true,
-  TODOComponent: true,
-  DONEComponent: true,
   videoComponent: true,
   audioComponent: true,
   pdfComponent: true,
-  blockquoteComponent: true,
-  tableComponent: true,
-  attributeComponent: true,
 }
-
-const orRegex = new RegExp(/{{or:(.*?)}}/, "g")
-const TODORegex = new RegExp(/{{.*?\bTODO\b.*?}}/, "g")
-const DONERegex = new RegExp(/{{.*?\bDONE\b.*?}}/, "g")
-const videoRegex = new RegExp(/{{.*?\[\[video\]\].*?\:(.*?)}}/, "g")
-const youtubeRegex = new RegExp(
-  /{{.*?\[\[video\]\].*?(https?:\/\/(?:www\.)?youtu(?:be\.com\/watch\?v=|\.be\/)([\w\-\_]*)(&(amp;)?[\w\?=]*)?)}}/,
-  "g",
-)
-
-// const multimediaRegex = new RegExp(/{{.*?\b(video|audio)\b.*?\:(.*?)}}/, "g")
-
-const audioRegex = new RegExp(/{{.*?\[\[audio\]\].*?\:(.*?)}}/, "g")
-const pdfRegex = new RegExp(/{{.*?\[\[pdf\]\].*?\:(.*?)}}/, "g")
-const blockquoteRegex = new RegExp(/(\[\[>\]\])\s*(.*)/, "g")
-const roamHighlightRegex = new RegExp(/\^\^(.+)\^\^/, "g")
-const roamItalicRegex = new RegExp(/__(.+)__/, "g")
-const tableRegex = new RegExp(/- {{.*?\btable\b.*?}}/, "g") /* TODO */
-const attributeRegex = new RegExp(/\b\w+(?:\s+\w+)*::/, "g") /* TODO */
 
 function isSpecialEmbed(node: Paragraph): boolean {
   if (node.children.length !== 2) return false
@@ -93,7 +61,7 @@ function transformSpecialEmbed(node: Paragraph, opts: Options): Html | null {
 
         return {
           type: "html",
-          value: `<iframe 
+          value: `<iframe
             class="external-embed youtube"
             width="600px"
             height="350px"
@@ -124,18 +92,18 @@ function transformSpecialEmbed(node: Paragraph, opts: Options): Html | null {
   }
 }
 
-export const RoamFlavoredMarkdown: QuartzTransformerPlugin<Partial<Options> | undefined> = (
+export const RoamFlavoredMarkdownSpecialEmbeds: MarkdownTransformerPlugin<Partial<Options>> = (
   userOpts,
 ) => {
   const opts = { ...defaultOptions, ...userOpts }
 
   return {
-    name: "RoamFlavoredMarkdown",
-    markdownPlugins() {
+    name: "RoamFlavoredMarkdownSpecialEmbeds",
+    transformation() {
       const plugins: PluggableList = []
 
       plugins.push(() => {
-        return (tree: Root, file: VFile) => {
+        return (tree: Root, _file: VFile) => {
           const replacements: [RegExp, ReplaceFunction][] = []
 
           // Handle special embeds (audio, video, PDF)
@@ -148,70 +116,6 @@ export const RoamFlavoredMarkdown: QuartzTransformerPlugin<Partial<Options> | un
                 }
               }
             }) as BuildVisitor<Root, "paragraph">)
-          }
-
-          // Roam italic syntax
-          replacements.push([
-            roamItalicRegex,
-            (_value: string, match: string) => ({
-              type: "emphasis",
-              children: [{ type: "text", value: match }],
-            }),
-          ])
-
-          // Roam highlight syntax
-          replacements.push([
-            roamHighlightRegex,
-            (_value: string, inner: string) => ({
-              type: "html",
-              value: `<span class="text-highlight">${inner}</span>`,
-            }),
-          ])
-
-          if (opts.orComponent) {
-            replacements.push([
-              orRegex,
-              (match: string) => {
-                const matchResult = match.match(/{{or:(.*?)}}/)
-                if (matchResult === null) {
-                  return { type: "html", value: "" }
-                }
-                const optionsString: string = matchResult[1]
-                const options: string[] = optionsString.split("|")
-                const selectHtml: string = `<select>${options.map((option: string) => `<option value="${option}">${option}</option>`).join("")}</select>`
-                return { type: "html", value: selectHtml }
-              },
-            ])
-          }
-
-          if (opts.TODOComponent) {
-            replacements.push([
-              TODORegex,
-              () => ({
-                type: "html",
-                value: `<input type="checkbox" disabled>`,
-              }),
-            ])
-          }
-
-          if (opts.DONEComponent) {
-            replacements.push([
-              DONERegex,
-              () => ({
-                type: "html",
-                value: `<input type="checkbox" checked disabled>`,
-              }),
-            ])
-          }
-
-          if (opts.blockquoteComponent) {
-            replacements.push([
-              blockquoteRegex,
-              (_match: string, _marker: string, content: string) => ({
-                type: "html",
-                value: `<blockquote>${content.trim()}</blockquote>`,
-              }),
-            ])
           }
 
           mdastFindReplace(tree, replacements)
