+++ quartz/plugins/transformers/resources/latex.ts
@@ -1,33 +1,14 @@
-import remarkMath from "remark-math"
-import rehypeKatex from "rehype-katex"
-import rehypeMathjax from "rehype-mathjax/svg"
-import { QuartzTransformerPlugin } from "../types"
+import { ExternalResourcePlugin } from "../../types"
 
 interface Options {
   renderEngine: "katex" | "mathjax"
-  customMacros: MacroType
 }
 
-interface MacroType {
-  [key: string]: string
-}
-
-export const Latex: QuartzTransformerPlugin<Partial<Options>> = (opts) => {
+export const Latex: ExternalResourcePlugin<Partial<Options>> = (opts) => {
   const engine = opts?.renderEngine ?? "katex"
-  const macros = opts?.customMacros ?? {}
   return {
     name: "Latex",
-    markdownPlugins() {
-      return [remarkMath]
-    },
-    htmlPlugins() {
-      if (engine === "katex") {
-        return [[rehypeKatex, { output: "html", macros }]]
-      } else {
-        return [[rehypeMathjax, { macros }]]
-      }
-    },
-    externalResources() {
+    transformation() {
       if (engine === "katex") {
         return {
           css: [
