+++ quartz/plugins/types.ts
@@ -6,22 +6,57 @@ import { FilePath } from "../util/path"
 import { BuildCtx } from "../util/ctx"
 import DepGraph from "../depgraph"
 
+type OptionType = object | undefined
 export interface PluginTypes {
-  transformers: QuartzTransformerPluginInstance[]
+  transformers: QuartzTransformerPluginInstance
   filters: QuartzFilterPluginInstance[]
   emitters: QuartzEmitterPluginInstance[]
 }
 
-type OptionType = object | undefined
-export type QuartzTransformerPlugin<Options extends OptionType = undefined> = (
+export type TextTransformerPlugin<Options extends OptionType = undefined> = (
   opts?: Options,
+) => TextTransformerPluginInstance
+export type TextTransformerPluginInstance = {
+  name: string
+  transformation: (ctx: BuildCtx, src: string | Buffer) => string | Buffer
+}
+
+export type MarkdownTransformerPlugin<Options extends OptionType = undefined> = (
+  opts?: Options,
+) => MarkdownTransformerPluginInstance
+export type MarkdownTransformerPluginInstance = {
+  name: string
+  transformation: (ctx: BuildCtx) => PluggableList
+}
+
+export type HtmlTransformerPlugin<Options extends OptionType = undefined> = (
+  opts?: Options,
+) => HtmlTransformerPluginInstance
+export type HtmlTransformerPluginInstance = {
+  name: string
+  transformation: (ctx: BuildCtx) => PluggableList
+}
+
+export type ExternalResourcePlugin<Options extends OptionType = undefined> = (
+  opts?: Options,
+) => ExternalResourcePluginInstance
+export type ExternalResourcePluginInstance = {
+  name: string
+  transformation: (ctx: BuildCtx) => Partial<StaticResources>
+}
+
+export type QuartzTransformerPlugin = (
+  textTransformers: TextTransformerPlugin,
+  markdownTransformers: MarkdownTransformerPlugin,
+  htmlTransformers: HtmlTransformerPlugin,
+  externalResources: ExternalResourcePlugin,
 ) => QuartzTransformerPluginInstance
 export type QuartzTransformerPluginInstance = {
-  name: string
-  textTransform?: (ctx: BuildCtx, src: string | Buffer) => string | Buffer
-  markdownPlugins?: (ctx: BuildCtx) => PluggableList
-  htmlPlugins?: (ctx: BuildCtx) => PluggableList
-  externalResources?: (ctx: BuildCtx) => Partial<StaticResources>
+  //name: string,
+  textTransformers: TextTransformerPluginInstance[]
+  markdownTransformers: MarkdownTransformerPluginInstance[]
+  htmlTransformers: HtmlTransformerPluginInstance[]
+  externalResources: ExternalResourcePluginInstance[]
 }
 
 export type QuartzFilterPlugin<Options extends OptionType = undefined> = (
